[refactor] Separate service and scheduler logic
