+++ backend/src/main/java/com/madeg/logistics/controller/RestoreController.java
@@ -2,7 +2,8 @@ package com.madeg.logistics.controller;
 
 import com.madeg.logistics.domain.CommonRes;
 import com.madeg.logistics.enums.ResponseCode;
-import com.madeg.logistics.scheduler.DatabaseBackupScheduler;
+import com.madeg.logistics.service.RestoreService;
+
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -16,24 +17,24 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
-@Tag(name = "Backup")
+@Tag(name = "Restore Database")
 @RestController
-@RequestMapping(path = "/api/v1/backup")
-public class BackupController {
+@RequestMapping(path = "/api/v1/restore")
+public class RestoreController {
 
     @Autowired
-    private DatabaseBackupScheduler databaseBackupScheduler;
+    private RestoreService restoreService;
 
-    @Operation(summary = "Restore Database")
+    @Operation(summary = "Restore Database using sql file")
     @ApiResponse(content = @Content(schema = @Schema(implementation = CommonRes.class)))
     @PostMapping
     public ResponseEntity<Object> restoreData(
             @RequestParam String fileName) {
         try {
-            databaseBackupScheduler.restoreDatabase(fileName);
+            restoreService.restoreDatabase(fileName);
             return ResponseEntity
                     .status(ResponseCode.SUCCESS.getStatus())
-                    .body(new CommonRes(ResponseCode.SUCCESS.getCode(), ResponseCode.SUCCESS.getMessage("DB restore")));
+                    .body(new CommonRes(ResponseCode.SUCCESS.getCode(), ResponseCode.SUCCESS.getMessage("DB 복구")));
         } catch (ResponseStatusException ex) {
             return ResponseEntity
                     .status(ex.getStatusCode())
